refactor(submission-gate): extract decision/retry helpers into decisions.ts
Move the cohesive cluster of pure decision/retry helper functions,
their RETRY_* constants, supporting timing constants, the
SourceEvidenceRetryableError data class, and the ReviewTarget /
DirectContentScope types out of the 5,137-line index.ts into a new
leaf module apps/submission-gate/src/decisions.ts, then import them
back. index.ts shrinks from 5,137 to 4,543 lines.

Only pure, side-effect-free functions/constants moved: decisionStatus,
gateCheckStatus, checksForDecision, decisionWithReviewContext,
decisionMetadata, nextReviewFor*, retryDelayFor*, isTimeoutError,
retryable*Decision, normalizeOneShotDecision, retry* fingerprint/
budget/backoff/state/exhausted helpers, *EvidenceConflict* and
*Contradicted helpers, plus isoAfter/isoBefore/nowIso/truncateForQueue.
All route handlers, the fetch handler, the queue consumer, and every
env/binding reader stay in index.ts. decisions.ts imports only from
sibling leaf modules (constants/review/github/duplicates/source-
evidence) and never from index.ts, so there is no import cycle.

This also resolves a latent unimported-type reference: the moved
evidence helpers now correctly import GateDecisionEvidence from
./review.

Source-pinned tests read index.ts and assert source.toContain(...) on
moved symbols, so readWorkerSource() in
tests/submission-gate-worker.test.ts now returns the concatenation of
index.ts + decisions.ts; every existing assertion still matches and
none were weakened.

Validation: pnpm type-check (clean); pnpm exec vitest run for
submission-gate-worker, private-reviewer-boundary, classify-pr-changes,
cleanup-policy, api-router-security (139 passed) plus submission-pr-first
(25 passed); prettier --check (clean); git diff --check (clean).
Behavior-preserving refactor with no runtime changes.

Closes #2262
